fix: types.ts to export interface
