+++ src/components/ApplicationBanner/types.ts
@@ -1,0 +1,8 @@
+export interface IAppAlert {
+  type: 'success' | 'warning' | 'info'
+  key: string
+  message?: string
+  canClose?: boolean
+  isHidden?: boolean
+  Component?: React.ElementType<{ className?: string }>
+}
